move output default to its own file now that its not part of the descriptor interface
